refactor(core): extract block tree entry to block_tree_entry.rs
Move BlockTreeEntry to dedicated block_tree_entry.rs module.
