+++ src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod block_tree_entry;
+
+pub use block_tree_entry::BlockTreeEntry;
